moved tests to separate file to avoid mypy crashes
